+++ src/components/users/register/StrengthPassword.ts
