+++ src/components/users/register/register.module.css
@@ -1,24 +1,3 @@
-.register-form{
-  display: flex;
-  flex-direction: column;
-  gap: 22px;
-  margin-bottom: 10px;
-  align-items: center;
-}
-
-.input{
-  position: relative;
-  padding: 2px;
-  margin: 18px 0;
-  vertical-align: middle;
-  width: 50%;
-}
-
-.paper{
-  padding: 24px 16px 16px;
-  margin: 0 -30%;
-}
-
 .register-avatars{
   max-width: 50%;
   flex-direction: row !important;
