+++ src/test/MainRoutesComponents.test.tsx
@@ -5,8 +5,8 @@ import { CreatorEditor } from "../components/creator/Editor/Editor"
 import { ActorSelection } from "../components/creator/ActorSelection/ActorSelection"
 import { Home } from "../components/home/Home"
 import { ImportedChallengeView } from "../components/ImportedChallengeView"
-import { PasswordRecovery } from "../components/PasswordRecovery"
-import { Register } from "../components/header/register/Register"
+import { PasswordRecovery } from "../components/users/passwordRecovery/PasswordRecovery"
+import { Register } from "../components/users/register/Register"
 import { renderComponent } from "./testUtils"
 import { validChallenge } from "./serializedChallenge.test"
 test('Renders home without errors', async () => {
